fix: import and export + tailwinf postcss configs
