+++ apps/deploy-web/src/pages/_app.tsx
@@ -30,7 +30,7 @@ import { queryClient } from "@src/queries";
 import { cn } from "@src/utils/styleUtils";
 
 import "nprogress/nprogress.css"; //styles of nprogress
-import "@akashnetwork/ui/styles/globals.css";
+import "@akashnetwork/ui/styles";
 import "../styles/index.css";
 import "@leapwallet/elements/styles.css";
