+++ apps/stats-web/package.json
@@ -56,7 +56,6 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@akashnetwork/tailwind-config": "*",
     "@types/json2csv": "^5.0.7",
     "@types/node": "20.14.0",
     "@types/react": "18.2.0",
