--- packages/tailwind-config/postcss.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  plugins: {
-    "postcss-import": {},
-    "tailwindcss/nesting": "postcss-nesting",
-    tailwindcss: {},
-    autoprefixer: {}
-  }
-};
