+++ packages/ui/package.json
@@ -8,15 +8,34 @@
   "author": "",
   "license": "ISC",
   "description": "",
+  "exports": {
+    "./components": "./components/index.tsx",
+    "./utils": "./utils/index.ts",
+    "./tailwind": "./tailwind.config.ts",
+    "./postcss": "./postcss.config.js",
+    "./styles": "./styles/global.css"
+  },
   "dependencies": {
     "@radix-ui/react-slot": "^1.0.2",
     "class-variance-authority": "0.7.0",
     "clsx": "2.1.1",
     "react": "18.2.0",
+    "react-dom": "18.2.0",
     "tailwind-merge": "2.3.0"
   },
   "devDependencies": {
+    "@akashnetwork/dev-config": "*",
+    "@tailwindcss/typography": "^0.5.13",
     "@types/react": "18.2.0",
-    "@types/react-dom": "18.2.0"
+    "@types/react-dom": "18.2.0",
+    "autoprefixer": "^10.4.16",
+    "prettier": "^3.3.0",
+    "prettier-plugin-tailwindcss": "^0.6.1",
+    "prop-types": "^15.8.1",
+    "tailwind-scrollbar": "^3.1.0",
+    "tailwindcss": "^3.4.3",
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "5.1.3"
   }
 }
+
