+++ packages/ui/postcss.config.js
@@ -1,1 +1,8 @@
-module.exports = require('../tailwind-config/postcss.config')
+module.exports = {
+  plugins: {
+    "postcss-import": {},
+    "tailwindcss/nesting": "postcss-nesting",
+    tailwindcss: {},
+    autoprefixer: {}
+  }
+};
